+++ NEWS.md
@@ -12,6 +12,8 @@
 ## New Functions
 
 - dist.point: distances to a given point;
+- centres.pp.roll: centres computed using a rolling window over a peptide chain;
+
 
 ## Updates to Existing Functions